Repositories are corrected

+++ src/main/java/project/repositories/PersonRepository.java
@@ -8,7 +8,7 @@ import project.models.Person;
 import java.util.Optional;
 
 @Repository
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 
      @Query("Select p from Person p where p.email = :email")
      Optional<Person> findPersonByEmail(String email);
